[breaking][hma] 1/n: Use signal_type_mapping instead of static signal / content types (#992)
* [hma] Add HMA version of pytx mappings classes

Test to verify that we're actually creating configs in dynamodb.

* [hma][pytx source] Allow switching to local pytx first time without failing

* Add an enforce method for getting signal and content types

Using this enforce method, replace usages of `threatexchange.content_type.meta.*` functions in hmalib.common.models.

* Correct typo in mappings method name

* Change workflow to do local builds for a while (until we release current pytx)

* Fix import and mypy complaints re pytx-alpha

* More typing fixes

* Write configs for HMASignalTypeMapping only when asked

- add method HMASignalTypeMapping.write_as_config which persists current state
- use SignalTypeMapping's constructor and provide classmethod to get config driven value
- Update all tests to use a secondary mechasnism to get a default instance of HMASignalTypeMapping

* Guard clause in set_threatexchange_source when deleting existing local-threatexchange

* Prevent work at module level in api_root

Re-do all subApp initializations inside a memoized bottle_init_once function.

* [minor][hmacli] do not fail if terraform outputs do not exist

* review: remove accidental imports and correctly place a docstring

* tiny: spread bottle_init_once results instead of subscripting

* Remove default signal/content type mappings. We are migrating without hmalib

* Use a tf module to run custom migrations as part of a release

This allows terraform to be the sole arbiter of HMA deployed versions.
Terraform will only run the migration if the migration has never been
run before or has been manually set as tainted by the user.

Setting it as tainted is a call taken by the user.

Test Plan
---
1. Verify that terraform apply runs the migration
```python
from hmalib.common.mappings import get_pytx_functionality_mapping
from hmalib.common.config import HMAConfig
HMAConfig.initialize('dipanjanm-HMAConfig')
print(get_pytx_functionality_mapping().signal_and_content)
```

Running it before terraform apply resulted in an empty dict.

After terraform apply

```python
>>> print(get_pytx_functionality_mapping().signal_and_content.content_by_name)
{'photo': <class 'threatexchange.content_type.photo.PhotoContent'>, 'video': <class 'threatexchange.content_type.video.VideoContent'>}
```

2. Verify that migration is not run in a second terraform apply

Done! Made some changes to the configs via dynamodb console. No changes
were reported.

* HMASignalTypeMapping no longer sub-class of SignalTypeMapping

@dcallies says no pytx component will 'require' signal_type_mapping be passed in. If we end up customizing it a lot, it is ok to not depend on the base class at all.

Remove tests because defaults are not done by code, but by a migration instead.

* review comments

* black fix

+++ hasher-matcher-actioner/hmalib/scripts/cli/main.py
@@ -20,7 +20,7 @@ import hmalib.scripts.cli.command_base as base
 import hmalib.scripts.cli.soak as soak
 import hmalib.scripts.cli.storm as storm
 import hmalib.scripts.cli.shell as shell
-from hmalib.scripts.cli import run_api, run_lambda, print_tfvars_example
+from hmalib.scripts.cli import run_api, run_lambda, print_tfvars_example, migrate
 
 TERRAFORM_OUTPUTS_CACHE = "/tmp/hma-terraform-outputs.json"
 
@@ -33,6 +33,7 @@ def get_subcommands() -> t.List[t.Type[base.Command]]:
         run_lambda.RunLambdaCommand,
         run_api.RunAPICommand,
         print_tfvars_example.PrintTFVarsExampleCommand,
+        migrate.MigrateCommand,
     ]
 
 
@@ -67,7 +68,7 @@ def get_argparse() -> argparse.ArgumentParser:
 
 
 def get_terraform_outputs(refresh=False) -> t.Dict:
-    if refresh:
+    if refresh and pathlib.Path(TERRAFORM_OUTPUTS_CACHE).exists():
         os.remove(TERRAFORM_OUTPUTS_CACHE)
 
     return utils.get_cached_terraform_outputs(TERRAFORM_OUTPUTS_CACHE)
